project time line and other issues fixed

+++ networklayer/src/main/java/com/emproto/networklayer/response/investment/Values.kt
@@ -4,7 +4,7 @@ data class Values(
     val dateOfCompletion: String,
     val displayName: String,
     val medias: Media,
-    val percentage: Int,
+    val percentage: Double,
     val status: String,
     val toolTipDetails: String,
     val webLink: String
